remove position.
