+++ sharding-jdbc/src/test/java/io/shardingsphere/dbtest/cases/authority/sql/SQLType.java
@@ -15,23 +15,18 @@
  * </p>
  */
 
-package io.shardingsphere.dbtest.cases.authority;
+package io.shardingsphere.dbtest.cases.authority.sql;
 
 /**
- * Authority SQL xml entry.
+ * SQL type for authority.
  *
  * @author panjuan
  */
-import lombok.Getter;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlAttribute;
-
-@Getter
-@XmlAccessorType(XmlAccessType.FIELD)
-public final class SQL {
+public enum SQLType {
     
-    @XmlAttribute(name = "content")
-    private String content;
+    Init("init"), Clean("clean");
+    
+    private final String sqlTypeName;
+    
+    SQLType(final String sqlTypeName) {this.sqlTypeName = sqlTypeName;}
 }
